+++ code/packages/rust/twig-to-semantic-ir/src/lower.rs
@@ -358,6 +358,7 @@ impl Lowerer {
             params: lam.params.iter().map(|p| Param {
                 name: p.clone(),
                 sir_type: None,
+                kind: semantic_ir::ParamKind::Required,
                 span: self.span_at(lam.line, lam.column),
             }).collect(),
             return_type: None,
@@ -627,6 +628,7 @@ impl Lowerer {
             params: lam.params.iter().map(|p| Param {
                 name: p.clone(),
                 sir_type: None,
+                kind: semantic_ir::ParamKind::Required,
                 span: span.clone(),
             }).collect(),
             return_type: None,